Rename translate/src/utils/message/serializer.ts -> serialize.ts
